feat(async dispose): allow asynchronous `dispose`
This feature allows for `dipose` of your dependency to return a promise.
Calling `dispose` on it's injector will await this promise.
This in turn makes `myInjector.dispose` asynchronous and you should await it.

BREAKING CHANGE: Dependencies are now disposed in asynchronously (while still honoring the order of "childFirst". You should now `await` the result of `injector.dispose()`.